+++ NEWS.md
@@ -1,10 +1,13 @@
 # withr (development version)
 
+- `with_environment()` now passing `pos` and `warn.conflicts` to `attach()`, as intended (#161).
+
+- `with_seed()` now also sets the RNG via new arguments `.rng_kind`, `.rng_normal_kind` and `.rng_sample_kind`
+  (#162, @AshesITR).
+
 - `with_timezone()` now works after recent changes to `Sys.timezone()` in R-devel (#165)
 
-- `with_environment()` now passing `pos` and `warn.conflicts` to `attach()`, as intended (#161)
-
-# withr 2.4.1
+withr 2.4.1
 
 - Tests which require `capabilities("cairo")` are now skipped.